// src/utils/enums/filter.ts
export enum FilterEnums {
    bank = 'bank',
    bank_type = 'bank_type',
    balanceType = 'balanceType',
    external_ref = 'external_ref',
    bank_account = 'bank_account',
    is_claimed = 'is_claimed',
    status = 'status',
    payout_status = 'payout_status',
    transactionId = 'transactionId',
    transactionIds = 'transactionIds',
    logger_status = 'logger_status',
    check_by_getaway = 'check_by_getaway',
    payment_action = 'payment_action', // Payin or Payout
    amount = 'amount',
    utr = 'utr',
    setting = 'setting',
    ifsc = 'ifsc',
    vendor = 'vendor',
    vendor_id = 'vendorId',
    reference_number = 'reference_number',
    transaction_id = 'transaction_id',
    date_range = 'date_range',
    // PAYOUT FILTERS
    order_id = 'order_id',
    account_number = 'account_number',
    has_vendor = 'has_vendor',
    payout_export_format = 'payout_export_format',
    batch_name = 'batch_name',
    balance_type = 'balance_type',
    noRelatedTo = 'noRelatedTo',
    categories = 'categories',
    bank_name = "bank_name",
    external_ref_id = "external_ref_id",
    balancetype = 'balancetype',
    gateway = "gateway",
    account_type = "accountType",
    channel_name = "channel_name",
    account_status = "account_status",
    ticket_status = "ticket_status",
    sender_account="sender_account",
    receiver_account = "receiver_account",
    transfer_from = 'transfer_from',
    transfer_to = 'transfer_to',
    user_type = 'user_type',
    transfer_type = 'transfer_type',
}