// src/utils/enums/role.ts
export enum RoleEnum {
    Admin = 'Admin',
    AccountLoginDetails = 'AccountLoginDetails',
    AccountDetails = 'AccountDetails',
    AccountPersonalDetails = 'AccountPersonalDetails',
    ActiveAllInProgressReport = 'Active All In Progress Report',
    ActiveReportLogs = 'Active Report Logs',
    AddCommentBatch = 'AddCommentBatch',
    AddSettlement = 'AddSettlement',
    AllowBulkUpdatePayoutTransactionUTR = 'Allow Bulk Update Payout Transaction UTR',
    AllowBulkUpdateTransactionUTR = 'Allow Bulk Update Transaction UTR',
    AllowBulkPayoutFailed = 'Allow Bulk Payout Failed',
    AnalyticsAccountTotal = 'AnalyticsAccountTotal',
    ApproveOrRejectTicket = 'ApproveOrRejectTicket',
    BalanceControl = 'BalanceControl',
    BalanceHistory = 'BalanceHistory',
    BalanceReport = 'BalanceReport',
    Bank = 'Bank',
    BankAccounts = 'Bank Accounts',
    BankAccountActivate = 'BankAccountActivate',
    BankAccountAmountRange = 'Bank Account Change Amount Range',
    BankAccountSetAmountRanges = 'Bank Account Update Amount Range Value',
    BankAccountChangePaymentMethod = 'Bank Account Change Payment Method',
    BankAccountClosingBalance = 'BankAccountClosingBalance',
    BankAccountCreate = 'BankAccountCreate',
    BankAccountCustomUpdate = 'BankAccountCustomUpdate',
    BankAccountStatusUpdate = 'BankAccountStatusUpdate',
    BankAccountUpdate = 'BankAccountUpdate',
    BankAccountUpdateKeyValue = 'Bank Account Update Key Value',
    Batch = 'Batch',
    Callback = 'Callback',
    CanExportUnlimitedBatches = 'Can Export Unlimited Batches',
    Client = 'Client',
    ControlBalanceHistory = 'ControlBalanceHistory',
    ControlBankTransaction = 'ControlBankTransaction',
    CreatePayoutTransactions = 'CreatePayoutTransactions',
    CreateTicket = 'CreateTicket', 
    ChargeBackInternalTransfer = 'Chargeback Internal Transfer',
    ClientChargeBack = 'Client Chargeback',
    ClientInternalTransfer = 'Client Internal Transfer',
    ClientTopUp = 'Client TopUp',
    Dashboard = 'Dashboard',
    Delete = 'Delete',
    EditAccountOtherFields = 'EditAccountOtherFields',
    EditPayoutTransactions = 'EditPayoutTransactions',
    EditPayoutTransactionsStatus = 'EditPayoutTransactionsStatus',
    EditPayoutTransactionsVendor = 'EditPayoutTransactionsVendor',
    EditTransactions = 'EditTransactions',
    Earning = 'Earning',
    EarningTotal = 'EarningTotal',
    ExceptProcessingAccounts = 'ExceptProcessingAccounts',
    FileAccess = 'FileAccess',
    GatewayAccounts = 'Gateway Accounts',
    HideBankAccountDetails = 'HideBankAccountDetails',
    HideSettingDetails = 'HideSettingDetails',
    ManualCheck = 'ManualCheck',
    MultiQRCodeAdd = 'Multi QR Code Adding',
    MultiQRCodeEdit = 'Multi QR Code Editing',
    MarkQRActiveOrInactive = 'Mark QR Active Or Inactive',
    OrdinaryUser = 'Transaction Report Without Client',
    PayinAccountsOnly = 'PayinAccountsOnly',
    PaymentPage = 'PaymentPage',
    PayoutAccountsOnly = 'PayoutAccountsOnly',
    PayoutDashboard = 'PayoutDashboard',
    PayoutTransactionReport = 'PayoutTransactionReport',
    PayoutTransactions = 'PayoutTransactions',
    ProcessingAccountsOnly = 'ProcessingAccountsOnly',
    ReconReport = 'ReconReport',
    RelatedTransaction = 'RelatedTransaction',
    Reports = 'Report Management',
    ResendCallback = 'ResendCallback',
    SendPayoutTransactionsToGateway = 'SendPayoutTransactionsToGateway',
    Settings = 'Settings',
    SettingsAnalytics = 'Settings Analytics',
    StatementManagement = 'Statement Management',
    StatementRecords = 'Statement Records',
    StatementReport = 'StatementReport',
    StatusExceptReadyToLive = 'StatusExceptReadyToLive',
    StatusHoldLiveOthers = 'StatusHoldLiveOthers',
    ShowDashboardSettlement = 'Show Dashboard Settlement',
    TicketSummary = 'TicketSummary',
    TransactionReport = 'TransactionReport',
    TransactionReportNew = 'TransactionReportNew',
    Transactions = 'Transactions',
    UnclaimedRecords = 'Unclaimed Records',
    UploadPayoutTransactionsFile = 'UploadPayoutTransactionsFile',
    UserControl = 'UserControl',
    VendorReport = 'VendorReport',
    ViewAccountOtherFields = 'ViewAccountOtherFields',
    ViewTicketSummaryData = 'ViewTicketSummaryData',
    TransferHistory = 'Transfer History',

}